move some stuff to a separate util file, fix gathering selection on open, always sync webview when command is invoked.
